+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [JKNIG-5] Add unit test for controller layer
+- Added unit tests for BookController in `BookControllerTest.java`
+
 ## [JKNIG-4] Add Spring Security feature
 - Added `spring-boot-starter-security` to pom.xml
 - Configured basic authentication with in-memory user in `SecurityConfig.java`